token/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
livekit-api==1.0.7
PyJWT>=2.0.0